# Speed up clr