Main:
```
· Discovering benchmarks
· Running 2 total benchmarks (1 commits * 1 environments * 2 benchmarks)
[ 0.00%] ·· Benchmarking existing-py_Users_mlondschien_mambaforge_envs_ivmodels_bin_python
[25.00%] ··· Running (tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration--)..
[75.00%] ··· tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration                                                                      ok
[75.00%] ··· ====== ============== ================ ============== =======================
             --                                      data                                 
             ------ ----------------------------------------------------------------------
               n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
             ====== ============== ================ ============== =======================
              1000     451±5μs         72.4±2ms      1.09±0.01ms          16.7±0.1ms      
             ====== ============== ================ ============== =======================

[100.00%] ··· tests.Tests.time_conditional_likelihood_ratio_test_power_series                                                                               ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     449±1μs        1.90±0.01s     1.10±0.01ms         2.81±0.03ms      
              ====== ============== ================ ============== =======================
```

here:
```
[ 0.00%] ·· Benchmarking existing-py_Users_mlondschien_mambaforge_envs_ivmodels_bin_python
[50.00%] ··· Running (tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration--).
[100.00%] ··· tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration                                                                      ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     456±7μs         37.9±2ms      1.08±0.02ms         1.01±0.01ms      
              ====== ============== ================ ============== =======================
```

honestly less of a speed-up than expected.