@mlondschien 2024-06-07
now:
```
[ 0.00%] ·· Benchmarking existing-py_Users_mlondschien_mambaforge_envs_ivmodels_bin_python
[50.00%] ··· Running (tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration--).
[100.00%] ··· tests.Tests.time_conditional_likelihood_ratio_test_numerical_integration                                                                                                                                                             ok
[100.00%] ··· ====== ============== ================ ============== =======================
              --                                      data                                 
              ------ ----------------------------------------------------------------------
                n     (1, 1, 0, 0)   (100, 1, 4, 0)   (4, 2, 2, 2)   guggenberger12 (k=10) 
              ====== ============== ================ ============== =======================
               1000     449±5μs         36.8±1ms      1.11±0.01ms         1.01±0.04ms      
              ====== ============== ================ ============== =======================
```