@mlondschien 2024-06-07
Superseded by https://github.com/mlondschien/ivmodels/pull/80